refactor: mudando tipo do id para integer

+++ src/main/java/backend/apiestacionamento/repository/AddressesRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AddressesRepository extends JpaRepository<Addresses, Long> {
+public interface AddressesRepository extends JpaRepository<Addresses, Integer> {
 }
